# 结构设计规范知识库/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HarnessAsset": {
        "properties": {
          "kind": {
            "enum": [
              "page_image",
              "image"
            ],
            "title": "Kind",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "path",
          "url",
          "source_file"
        ],
        "title": "HarnessAsset",
        "type": "object"
      },
      "HarnessSearchRequest": {
        "additionalProperties": false,
        "description": "Read-only search contract consumed by external agent adapters.",
        "properties": {
          "document": {
            "default": "",
            "maxLength": 240,
            "title": "Document",
            "type": "string"
          },
          "include_assets": {
            "default": false,
            "title": "Include Assets",
            "type": "boolean"
          },
          "mode": {
            "default": "auto",
            "enum": [
              "auto",
              "table",
              "clause",
              "definition",
              "general"
            ],
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "HarnessSearchRequest",
        "type": "object"
      },
      "HarnessSearchResponse": {
        "properties": {
          "data_version_hash": {
            "title": "Data Version Hash",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "normalized_query": {
            "title": "Normalized Query",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/HarnessSearchResult"
            },
            "title": "Results",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "query",
          "normalized_query",
          "mode",
          "data_version_hash",
          "result_count",
          "results"
        ],
        "title": "HarnessSearchResponse",
        "type": "object"
      },
      "HarnessSearchResult": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/HarnessAsset"
            },
            "title": "Assets",
            "type": "array"
          },
          "authority_level": {
            "title": "Authority Level",
            "type": "integer"
          },
          "clause_number": {
            "title": "Clause Number",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "is_table": {
            "title": "Is Table",
            "type": "boolean"
          },
          "matched_terms": {
            "items": {
              "type": "string"
            },
            "title": "Matched Terms",
            "type": "array"
          },
          "pages": {
            "items": {
              "type": "integer"
            },
            "title": "Pages",
            "type": "array"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "retrieval_sources": {
            "items": {
              "type": "string"
            },
            "title": "Retrieval Sources",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "section_type": {
            "title": "Section Type",
            "type": "string"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "retrieval",
              "structured_table"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "standard_code": {
            "title": "Standard Code",
            "type": "string"
          },
          "standard_name": {
            "title": "Standard Name",
            "type": "string"
          },
          "structured_row": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Row"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          },
          "table_name": {
            "title": "Table Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "source_kind",
          "source_file",
          "standard_code",
          "standard_name",
          "version",
          "section_type",
          "authority_level",
          "is_table",
          "clause_number",
          "table_id",
          "table_name",
          "pages",
          "excerpt",
          "score",
          "reason"
        ],
        "title": "HarnessSearchResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "src__app__api__knowledge__CandidateStatusUpdate": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "CandidateStatusUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "结构设计规范知识库 RAG API (多模态)",
    "version": "3.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/corrections/candidates/{doc}": {
      "get": {
        "operationId": "correction_candidate_detail_corrections_candidates__doc__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc",
            "required": true,
            "schema": {
              "title": "Doc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correction Candidate Detail"
      }
    },
    "/corrections/candidates/{doc}/{candidate_id}": {
      "patch": {
        "operationId": "correction_candidate_update_corrections_candidates__doc___candidate_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "doc",
            "required": true,
            "schema": {
              "title": "Doc",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/src__app__api__knowledge__CandidateStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correction Candidate Update"
      }
    },
    "/corrections/promote/{doc}": {
      "post": {
        "operationId": "correction_promote_corrections_promote__doc__post",
        "parameters": [
          {
            "in": "path",
            "name": "doc",
            "required": true,
            "schema": {
              "title": "Doc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correction Promote"
      }
    },
...
    "/integrations/deepseek-harness/page": {
      "get": {
        "operationId": "harness_page_integrations_deepseek_harness_page_get",
        "parameters": [
          {
            "in": "query",
            "name": "source_file",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Source File",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Harness Page Integrations Deepseek Harness Page Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Harness Page",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/integrations/deepseek-harness/search": {
      "post": {
        "operationId": "harness_search_integrations_deepseek_harness_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarnessSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Harness Search",
        "tags": [
          "integrations"
        ]
      }
    },
...
  }
}